ai-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=9.5.0
baidu-aip>=4.16.0
langchain>=0.0.300
langchain-community>=0.0.300
python-multipart>=0.0.6